machining_feature_localizer/requirements.txt
wandb==0.15.12
optuna==3.4.0
numpy-stl==3.0.1
scikit-learn==1.3.2
thop==0.1.1.post2209072238
Flask==3.0.3
flask-cors==5.0.0
pymadcad==0.16.0
